+++ src/components/Common/Modals/ModalLayout.tsx
@@ -5,7 +5,7 @@ import {
   ModalActionProps,
 } from 'src/redux/reducers/components/modals';
 import { removeSigninErrorMsg } from 'src/redux/reducers/components/validation';
-import Logo from '../Logo';
+import Logo from '../Logos/Logo';
 import { Backdrop } from './Backdrop';
 
 const Container = styled.div`
